requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1

# HTTP + resilience
requests==2.31.0
tenacity==8.3.0
httpx==0.27.0

# LLM
groq==0.9.0

# Blockchain (Phase 4+)
py-algorand-sdk==2.6.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7

# Phase 5+ (uncomment when needed)
# pyteal==0.25.0
# algokit-utils==2.2.1
